Friends ID/Profiles detection/Server list info
Profile detection:
The entire steamcommunity URL to a profile can now be provided when
querying through Steam Friends Status.

Server list:
Add sorting (which doesn't work as it should on some columns since
it does string based sorting), and resize the columns to take up
as much space as possible.