Split Steam Community-centric queries to class
This commit is mainly a backend cleanup, dividing things
up where they should be instead of having one giant mess.

One minor change is that the timeout for urllib2 has been set to
4 seconds through the use of the socket module to avoid locking
the UI forever if Steam Community is unreachable (needs threading).

The class should eventually be released like hlquery as well.